HMatrices.jl

A package for assembling and factoring hierarchical matrices with a focus on boundary integral equations

This package provides various methods for assembling as well as for doing linear algebra with hierarchical matrices. The main structure exported is the HMatrix type, which can be used to efficiently approximate certain linear operators containing a hierarchical low-rank structure. Once assembled, a hierarchical matrix can be used to accelerate the solution of Ax=b in a variety of ways, as illustrated in the examples that follow.

Note

Although hierarchical matrices have a broad range of application, this package focuses on their use to approximate integral operators arising in boundary integral equation methods.

Assembling an HMatrix

In order to assemble an HMatrix, you need the following (problem-specific) ingredients:

The matrix-like object K that you wish to compress

A hierarchical partition of the rows and columns indices of K

An admissibility condition for determining (a priory) whether a block is compressible, and

A function/functor to generate a low-rank approximation of compressible blocks

To illustrate how this is done for a concrete problem, consider two set of points $X = \left\{ \boldsymbol{x}_i \right\}_{i=1}^m$ and $Y = \left\{ \boldsymbol{x}_j \right\}_{j=1}^m$ in $\mathbb{R}^3$, and let A be an $m \times n$ matrix with entries given by:

\[ A_{i,j} = \exp(-||\boldsymbol{x}_i-\boldsymbol{y}_j||)\]

The following code will create the point clouds X and Y, as well as the matrix-like object K:

using HMatrices, LinearAlgebra, StaticArrays
const Point3D = SVector{3,Float64}

struct ExponentialMatrix <: AbstractMatrix{Float64}
  X::Vector{Point3D}  
  Y::Vector{Point3D}
end

Base.getindex(K::ExponentialMatrix,i::Int,j::Int) = exp(-norm(K.X[i] - K.Y[j]))
Base.size(K::ExponentialMatrix) = length(K.X), length(K.Y)

# points on a sphere
m = n = 10000
X = Y = [Point3D(sin(θ)cos(ϕ),sin(θ)*sin(ϕ),cos(θ)) for (θ,ϕ) in zip(π*rand(n),2π*rand(n))]
# create the abstract matrix
K = ExponentialMatrix(X,Y)

The next step in to partition the point clouds X and Y into a tree-like data structure so that blocks corresponding to well-separated points can be easily distinguished and compressed. The WavePropBase package provides the ClusterTree struct for this purpose:

Xclt = Yclt = ClusterTree(X)
nothing # hide

The third requirement is an admissibilty condition to determine if the interaction between two clusters should be compressed. We will use the StrongAdmissibilityStd:

adm = StrongAdmissibilityStd()

The final step is to provide a method to compress admissible blocks. Here we will use the PartialACA functor:

comp = PartialACA(;atol=1e-6)

We can now compress the operator K using

H = HMatrix(K,Xclt,Yclt,adm,comp;threads=false,distributed=false)

Matrix vector product and iterative solvers

You can now use H as an approximation to K:

x = rand(n)
H*x ≈ K*x 

The larger the number of points n, the more we expect to gain from this compression technique.

Factorization and direct solvers

Although the forward map illustrated in the example above suffices to solve the linear system Kx = b, there are circumnstances when a direct solver is more appropriate (because, e.g., the system is not well-conditioned or you wish to solve it for many right-hand-sides b). At present, you can do an lu factorization of H as follows:

F = lu(H,comp)

The factorization F can then be used to solve a linear system:

b = rand(m)
approx = F\b
exact  = Matrix(K)\b
norm(approx-exact,Inf)

Note

Most algebraic manipulations of hierarchical matrices require the use of compressor to keep the rank of the low blocks under control when e.g. adding two low-rank blocks. In HMatrices.jl, there are a few options defined in src/compressor.jl.

Note that the error in solving the linear system may be significantly larger than the error in computing H*x due to the condition of the underlying operator.

Tip

Because factoring an HMatrix with a small error tolerance can be quite time-consuming, a hybrid strategy consists of using a rough factorization (with a large tolerance) as a preconditioner to an iterative solver.

Tensor-valued matrices

Some support is currently provided for matrices with entries which are themselves matrices...

References

[1] Hackbusch, Wolfgang. Hierarchical matrices: algorithms and analysis. Vol. 49. Heidelberg: Springer, 2015.

[2] Bebendorf, Mario. Hierarchical matrices. Springer Berlin Heidelberg, 2008.

Base.Matrix

HMatrices.ACA

HMatrices.AbstractCompressor

HMatrices.HMatrix

HMatrices.HMatrix

HMatrices.HMatrix

HMatrices.HMulNode

HMatrices.PartialACA

HMatrices.PermutedMatrix

HMatrices.RkMatrix

HMatrices.RkMatrix

HMatrices.RkMatrix

HMatrices.RkMatrix

HMatrices.StrongAdmissibilityStd

HMatrices.TSVD

HMatrices.WeakAdmissibilityStd

Base.hcat

Base.vcat

HMatrices._aca_full!

HMatrices._aca_full_pivot

HMatrices._aca_partial

HMatrices._aca_partial_pivot

HMatrices._build_block_structure!

HMatrices._hgemv_recursive!

HMatrices._update_frob_norm

HMatrices.assemble_cpu!

HMatrices.assemble_threads!

HMatrices.build_sequence_partition

HMatrices.compress!

HMatrices.compress!

HMatrices.compression_ratio

HMatrices.compression_ratio

HMatrices.cost_mv

HMatrices.find_optimal_cost

HMatrices.flush_to_children!

HMatrices.getcol

HMatrices.getcol!

HMatrices.has_partition

HMatrices.hilbert_cartesian_to_linear

HMatrices.hilbert_linear_to_cartesian

HMatrices.hilbert_partitioning

HMatrices.hmul!

HMatrices.isclean

HMatrices.num_stored_elements

LinearAlgebra.axpy!

HMatrices.@hprofile

Base.Matrix — Method

Matrix(H::HMatrix;global_index=false)

Convert H to a Matrix. If global_index=true, the entries are given in the global indexing system (see HMatrix for more information); otherwise the local indexing system induced by the row and columns trees are used (default).

HMatrices.ACA — Type

struct ACA

Adaptive cross approximation algorithm with full pivoting. This structure can be used to generate an RkMatrix from a matrix-like object M as follows:

using LinearAlgebra
rtol = 1e-6
comp = ACA(;rtol)
A = rand(10,2)
B = rand(10,2)
M = A*adjoint(B) # a low-rank matrix
R = comp(M,:,:) # compress the entire matrix `M`
norm(Matrix(R) - M) < rtol*norm(M) # true

# output

true

The keywork arguments rtol, atol, and rank can be used to control the quality of the approximation. Because it uses full pivoting, the linear operator has to be evaluated at every i,j. See also: [PartialACA](@ref).

HMatrices.AbstractCompressor — Type

abstract type AbstractCompressor

Types used to compress matrices.

HMatrices.HMatrix — Type

mutable struct HMatrix{R,T} <: AbstractHMatrix{T}

A hierarchial matrix constructed from a rowtree and coltree of type R and holding elements of type T.

HMatrices.HMatrix — Type

HMatrix(K,rowtree,coltree,adm,comp;threads=true,distributed=false,permute_kernel=true)

Main constructor for hierarchical matrix, where K represents the matrix to be approximated, rowtree and coltree are tree structure partitioning the row and column indices, respectively, adm can be called on a node of rowtree and a node of coltree to determine if the block is compressible, and comp is a function/functor which can compress admissible blocks.

It is assumed that K supports getindex(K,i,j), and that comp can be called as comp(K,irange::UnitRange,jrange::UnitRange) to produce a compressed version of K[irange,jrange].

HMatrices.HMatrix — Method

HMatrix{T}(rowtree,coltree,adm)

Construct an empty HMatrix with rowtree and coltree using the admissibility condition adm.

HMatrices.HMulNode — Type

struct HMulNode{S,T} <: AbstractTree

Tree data structure representing the following computation:

    C <-- C + a * ∑ᵢ Aᵢ * Bᵢ

where C = target(node), and Aᵢ,Bᵢ are pairs stored in sources(node).

This structure is used to group the operations required when multiplying hierarchical matrices so that they can later be executed in a way that minimizes recompression of intermediate computations.

HMatrices.PartialACA — Type

struct PartialACA

Adaptive cross approximation algorithm with partial pivoting. This structure can be used to generate an RkMatrix from a matrix-like object M as follows:

using LinearAlgebra
rtol = 1e-6
comp = PartialACA(;rtol)
A = rand(10,2)
B = rand(10,2)
M = A*adjoint(B) # a low-rank matrix
R = comp(M) # compress the entire matrix `M`
norm(Matrix(R) - M) < rtol*norm(M) # true

# output

true

Because it uses partial pivoting, the linear operator does not have to be evaluated at every i,j. This is usually much faster than ACA, but due to the pivoting strategy the algorithm may fail in special cases, even when the underlying linear operator is of low rank.

HMatrices.PermutedMatrix — Type

PermutedMatrix{K,T} <: AbstractMatrix{T}

Structured used to reprensent the permutation of a matrix-like object. The original matrix is stored in the data::K field, and the permutations are stored in rowperm and colperm.

HMatrices.RkMatrix — Type

mutable struct RkMatrix{T}

Representation of a rank r matrix M in outer product format M = A*adjoint(B) where A has size m × r and B has size n × r.

The internal representation stores A and B, but R.Bt or R.At can be used to get the respective adjoints.

HMatrices.RkMatrix — Method

RkMatrix(F::LinearAlgebra.SVD)

Construct an RkMatrix from an SVD factorization.

HMatrices.RkMatrix — Method

RkMatrix(M::Matrix)

Construct an RkMatrix from a Matrix by passing through the full svd of M.

HMatrices.RkMatrix — Method

RkMatrix(A::Vector{<:Vector},B::Vector{<:Vector})

Construct an RkMatrix from a vector of vectors. Assumes that length(A) == length(B), which determines the rank, and that all vectors in A (resp. B) have the same length m (resp. n).

HMatrices.StrongAdmissibilityStd — Type

struct StrongAdmissibilityStd

Two blocks are admissible under this condition if the minimum of their diameter is smaller than eta times the distance between them, where eta::Float64 is a parameter.

Usage:

adm = StrongAdmissibilityStd(;eta=2.0)
adm(Xnode,Ynode)

HMatrices.TSVD — Type

struct TSVD

Compression algorithm based on a posteriori truncation of an SVD. This is the optimal approximation in Frobenius norm; however, it also tends to be very expensive and thus should be used mostly for "small" matrices.

HMatrices.WeakAdmissibilityStd — Type

struct WeakAdmissibilityStd

Two blocks are admissible under this condition if the distance between them is positive.

Base.hcat — Method

hcat(M1::RkMatrix,M2::RkMatrix)

Concatenated M1 and M2 horizontally to produce a new RkMatrix of rank rank(M1)+rank(M2).

Base.vcat — Method

vcat(M1::RkMatrix,M2::RkMatrix)

Concatenated M1 and M2 vertically to produce a new RkMatrix of rank rank(M1)+rank(M2)

HMatrices._aca_full! — Method

_aca_full!(M,atol,rmax,rtol)

Internal function implementing the adaptive cross-approximation algorithm with full pivoting. The matrix M is modified in place. The returned RkMatrix has rank at most rmax, and is expected to satisfy |M - R| < max(atol,rtol*|M|).

HMatrices._aca_full_pivot — Method

_aca_full_pivot(M)

Find the index of the element x ∈ M maximizing its smallest singular value. This is equivalent to minimizing the spectral norm of the inverse of x.

When x is a scalar, this is simply the element with largest absolute value.

See also: _aca_partial_pivot.

HMatrices._aca_partial — Function

_aca_partial(K,irange,jrange,atol,rmax,rtol,istart=1)

Internal function implementing the adaptive cross-approximation algorithm with partial pivoting. The returned R::RkMatrix provides an approximation to K[irange,jrange] which has either rank is expected to satisfy|M - R| < max(atol,rtol*|M|)`, but this inequality may fail to hold due to the various errors involved in estimating the error and |M|.

HMatrices._aca_partial_pivot — Method

_aca_partial_pivot(v,I)

Find in the valid set I the index of the element x ∈ v maximizing its smallest singular value. This is equivalent to minimizing the spectral norm of the inverse of x.

When x is a scalar, this is simply the element with largest absolute value.

This general implementation should work for both scalar as well as tensor-valued kernels; see (https://www.sciencedirect.com/science/article/pii/S0021999117306721)[https://www.sciencedirect.com/science/article/pii/S0021999117306721] for more details.

HMatrices._build_block_structure! — Method

_build_block_structure!(adm_fun,current_node)

Recursive constructor for HMatrix block structure. Should not be called directly.

HMatrices._hgemv_recursive! — Method

_hgemv_recursive!(C,A,B,offset)

Internal function used to compute C[I] <-- C[I] + A*B[J] where I = rowrange(A) - offset[1] and J = rowrange(B) - offset[2].

The offset argument is used on the caller side to signal if the original hierarchical matrix had a pivot other than (1,1).

HMatrices._update_frob_norm — Method

_update_frob_norm(acc,A,B)

Given the Frobenius norm of Rₖ = A[1:end-1]*adjoint(B[1:end-1]) in acc, compute the Frobenius norm of Rₖ₊₁ = A*adjoint(B) efficiently.

HMatrices.assemble_cpu! — Method

assemble_cpu!(hmat::HMatrix,K,comp)

Assemble data on the leaves of hmat. The admissible leaves are compressed using the compressor comp. This function assumes the structure of hmat has already been intialized, and therefore should not be called directly. See HMatrix information on constructors.

HMatrices.assemble_threads! — Method

assemble_threads!(hmat::HMatrix,K,comp)

Like assemble_cpu!, but uses threads to assemble (independent) blocks.

HMatrices.build_sequence_partition — Method

build_sequence_partition(seq,nq,cost,nmax)

Partition the sequence seq into nq contiguous subsequences with a maximum of cost of nmax per set. Note that if nmax is too small, this may not be possible (see has_partition).

HMatrices.compress! — Method

compress!(M::RkMatrix,tsvd::TSVD)

Recompress the matrix R using a truncated svd of R. The implementation uses the qr-svd strategy to efficiently compute svd(R) when rank(R) ≪ min(size(R)).

HMatrices.compress! — Method

compress!(M::Matrix,tsvd::TSVD)

Recompress the matrix M using a truncated svd and output an RkMatrix. The data in M is invalidated in the process.

HMatrices.compression_ratio — Method

compression_ratio(R::RkMatrix)

The ratio of the uncompressed size of R to its compressed size in outer product format.

HMatrices.compression_ratio — Method

compression_ratio(H::HMatrix)

The ratio of the uncompressed size of H to its compressed size.

HMatrices.cost_mv — Method

cost_mv(A::Union{Matrix,SubArray,Adjoint})

A proxy for the computational cost of a matrix/vector product.

HMatrices.find_optimal_cost — Function

find_optimal_cost(seq,nq,cost,tol)

Find an approximation to the cost of an optimal partitioning of seq into nq contiguous subsequent. The optimal cost is the smallest number cmax such that has_partition(seq,nq,cost,cmax) returns true.

HMatrices.flush_to_children! — Method

flush_to_children!(H::HMatrix,compressor)

Transfer the blocks data to its children. At the end, set H.data to nothing.

HMatrices.getcol! — Method

getcol!(col,M::AbstractMatrix,j)

Fill the entries of col with column j of M.

HMatrices.getcol — Method

getcol(M::AbstractMatrix,j)

Return a vector containing the j-th column of M.

HMatrices.has_partition — Function

has_partition(v,np,cost,cmax)

Given a vector v, determine whether or not a partition into np segments is possible where the cost of each partition does not exceed cmax.

HMatrices.hilbert_cartesian_to_linear — Method

hilbert_cartesian_to_linear(n,x,y)

Convert the cartesian indices x,y into a linear index d using a hilbert curve of order n. The coordinates x,y range from 0 to n-1, and the output d ranges from 0 to n^2-1.

See https://en.wikipedia.org/wiki/Hilbert_curve.

HMatrices.hilbert_linear_to_cartesian — Method

hilbert_linear_to_cartesian(n,d)

Convert the linear index 0 ≤ d ≤ n^2-1 into the cartesian coordinates 0 ≤ x < n-1 and 0 ≤ y ≤ n-1 on the Hilbert curve of order n.

See https://en.wikipedia.org/wiki/Hilbert_curve.

HMatrices.hilbert_partitioning — Function

hilbert_partitioning(H::HMatrix,np,[cost=cost_mv])

Partiotion the leaves of H into np sequences of approximate equal cost (as determined by the cost function) while also trying to maximize the locality of each partition.

HMatrices.hmul! — Method

hmul!(C::HMatrix,A::HMatrix,B::HMatrix,a,b,compressor)

Similar to LinearAlgebra.mul! : compute C <-- A*B*a + B*b, where A,B,C are hierarchical matrices and compressor is a function/functor used in the intermediate stages of the multiplication to avoid growring the rank of admissible blocks after addition is performed.

HMatrices.isclean — Method

isclean(H::HMatrix)

Return true if H all leaves of H have data, and if the leaves are the only nodes containing data. This is the normal state of an ℋ-matrix, but during intermediate stages of a computation data may be associated with non-leaf nodes for convenience.

HMatrices.num_stored_elements — Method

num_stored_elements(R::RkMatrix)

The number of entries stored in the representation. Note that this is not length(R).

LinearAlgebra.axpy! — Method

LinearAlgebra.axpy!(a::Number,X::Union{Matrix,RkMatrix,HMatrix},Y::Union{Matrix,RkMatrix,HMatrix})

Perform Y <-- a*X + Y in-place. Note that depending on the types of X and Y, this may require converting from/to different formats during intermdiate calculations.

In the case where Y is an RkMatrix, the call axpy!(a,X,Y) should typically be followed by recompression stage to keep the rank of Y under control.

In the case where Y is an HMatrix, the call axpy!(a,X,Y) sums X to the data in the node Y (and not on the leaves). In case Y has no data, it will simply be assigned X. This means that after the call axpy(a,X,Y), the object Y is in a dirty state (see [isclean][@ref]) and usually a call to flush_to_leaves! or flush_to_children! follows.

HMatrices.@hprofile — Macro

@hprofile

A macro which

resets the default TimerOutputs.get_defaulttimer to zero

execute the code block

print the profiling details

This is useful as a coarse-grained profiling strategy in HMatrices to get a rough idea of where time is spent. Note that this relies on TimerOutputs annotations manually inserted in the code.