References

HMatrices.ALLOW_GETINDEX — Constant

const ALLOW_GETINDEX

If set to false, the getindex(H,i,j) method will throw an error on AbstractHMatrix.

Base.Matrix — Method

Matrix(H::HMatrix;global_index=false)

Convert H to a Matrix. If global_index=true, the entries are given in the global indexing system (see HMatrix for more information); otherwise the local indexing system induced by the row and columns trees are used (default).

HMatrices.ACA — Type

struct ACA

Adaptive cross approximation algorithm with full pivoting. This structure can be used to generate an RkMatrix from a matrix-like object M. The keywork arguments rtol, atol, and rank can be used to control the quality of the approximation. Note that because ACA uses full pivoting, the linear operator M has to be evaluated at every i,j.

See also: [PartialACA](@ref)

Examples

using LinearAlgebra
rtol = 1e-6
comp = ACA(;rtol)
A = rand(10,2)
B = rand(10,2)
M = A*adjoint(B) # a low-rank matrix
R = comp(M,:,:) # compress the entire matrix `M`
norm(Matrix(R) - M) < rtol*norm(M) # true

# output

true

HMatrices.AbstractCompressor — Type

abstract type AbstractCompressor

Types used to compress matrices.

HMatrices.AbstractKernelMatrix — Type

abstract type AbstractKernelMatrix{T} <: AbstractMatrix{T}

Interface for abstract matrices represented through a kernel function f, target elements X, and source elements Y. The matrix entry i,j is given by f(X[i],Y[j]). Concrete subtypes should implement at least

`Base.getindex(K::AbstractKernelMatrix,i::Int,j::Int)`

If a more efficient implementation of getindex(K,I::UnitRange,I::UnitRange), getindex(K,I::UnitRange,j::Int) and getindex(adjoint(K),I::UnitRange,j::Int) is available (e.g. with SIMD vectorization), implementing such methods can improve the speed of assembling an HMatrix.

HMatrices.DHMatrix — Type

mutable struct DHMatrix{R,T} <: AbstractHMatrix{T}

Concrete type representing a hierarchical matrix with data distributed amongst various workers. Its structure is very similar to HMatrix, except that the leaves store a RemoteHMatrix object.

The data on the leaves of a DHMatrix may live on a different worker, so calling fetch on them should be avoided whenever possible.

HMatrices.DHMatrix — Method

DHMatrix{T}(rowtree,coltree;partition_strategy=:distribute_columns)

Construct the block structure of a distributed hierarchical matrix covering rowtree and coltree. Returns a DHMatrix with leaves that are empty.

The partition_strategy keyword argument determines how to partition the blocks for distributed computing. Currently, the only available options is distribute_columns, which will partition the columns of the underlying matrix into floor(log2(nw)) parts, where nw is the number of workers available.

HMatrices.HMatrix — Type

mutable struct HMatrix{R,T} <: AbstractHMatrix{T}

A hierarchial matrix constructed from a rowtree and coltree of type R and holding elements of type T.

HMatrices.HMatrix — Method

HMatrix{T}(rowtree,coltree,adm)

Construct an empty HMatrix with rowtree and coltree using the admissibility condition adm. This function builds the skeleton for the hierarchical matrix, but does not compute data field in the blocks. See assemble_hmat for assembling a hierarhical matrix.

HMatrices.HMulNode — Type

struct HMulNode{S,T} <: AbstractTree

Tree data structure representing the following computation:

    C <-- C + a * ∑ᵢ Aᵢ * Bᵢ

where C = target(node), and Aᵢ,Bᵢ are pairs stored in sources(node).

This structure is used to group the operations required when multiplying hierarchical matrices so that they can later be executed in a way that minimizes recompression of intermediate computations.

HMatrices.KernelMatrix — Type

KernelMatrix{Tf,Tx,Ty,T} <:: AbstractKernelMatrix{T}

Generic kernel matrix representing a kernel function acting on two sets of elements. If K is a KernelMatrix, then K[i,j] = f(X[i],Y[j]) where f::Tf=kernel(K), X::Tx=rowelements(K) and Y::Ty=colelements(K).

Examples

X = rand(Geometry.Point2D,100)
Y = rand(Geometry.Point2D,100)
K = KernelMatrix(X,Y) do x,y
    sum(x+y)
end

HMatrices.PartialACA — Type

struct PartialACA

Adaptive cross approximation algorithm with partial pivoting. This structure can be used to generate an RkMatrix from a matrix-like object M as follows:

using LinearAlgebra
rtol = 1e-6
comp = PartialACA(;rtol)
A = rand(10,2)
B = rand(10,2)
M = A*adjoint(B) # a low-rank matrix
R = comp(M) # compress the entire matrix `M`
norm(Matrix(R) - M) < rtol*norm(M) # true

# output

true

Because it uses partial pivoting, the linear operator does not have to be evaluated at every i,j. This is usually much faster than ACA, but due to the pivoting strategy the algorithm may fail in special cases, even when the underlying linear operator is of low rank.

HMatrices.PermutedMatrix — Type

PermutedMatrix{K,T} <: AbstractMatrix{T}

Structured used to reprensent the permutation of a matrix-like object. The original matrix is stored in the data::K field, and the permutations are stored in rowperm and colperm.

HMatrices.RemoteHMatrix — Type

struct RemoteHMatrix{S,T}

A light wrapper for a Future storing an HMatrix.

HMatrices.RkMatrix — Type

mutable struct RkMatrix{T}

Representation of a rank r matrix M in outer product format M = A*adjoint(B) where A has size m × r and B has size n × r.

The internal representation stores A and B, but R.Bt or R.At can be used to get the respective adjoints.

HMatrices.RkMatrix — Method

RkMatrix(F::SVD)

Construct an RkMatrix from an SVD factorization.

HMatrices.RkMatrix — Method

RkMatrix(M::Matrix)

Construct an RkMatrix from a Matrix by passing through the full svd of M.

HMatrices.RkMatrix — Method

RkMatrix(A::Vector{<:Vector},B::Vector{<:Vector})

Construct an RkMatrix from a vector of vectors. Assumes that length(A) == length(B), which determines the rank, and that all vectors in A (resp. B) have the same length m (resp. n).

HMatrices.StrongAdmissibilityStd — Type

struct StrongAdmissibilityStd

Two blocks are admissible under this condition if the minimum of their diameter is smaller than eta times the distance between them, where eta::Float64 is a parameter.

Usage:

adm = StrongAdmissibilityStd(;eta=2.0)
adm(Xnode,Ynode)

HMatrices.TSVD — Type

struct TSVD

Compression algorithm based on a posteriori truncation of an SVD. This is the optimal approximation in Frobenius norm; however, it also tends to be very expensive and thus should be used mostly for "small" matrices.

HMatrices.WeakAdmissibilityStd — Type

struct WeakAdmissibilityStd

Two blocks are admissible under this condition if the distance between them is positive.

Base.hcat — Method

hcat(M1::RkMatrix,M2::RkMatrix)

Concatenated M1 and M2 horizontally to produce a new RkMatrix of rank rank(M1)+rank(M2).

Base.vcat — Method

vcat(M1::RkMatrix,M2::RkMatrix)

Concatenated M1 and M2 vertically to produce a new RkMatrix of rank rank(M1)+rank(M2)

HMatrices._aca_full! — Method

_aca_full!(M,atol,rmax,rtol)

Internal function implementing the adaptive cross-approximation algorithm with full pivoting. The matrix M is modified in place. The returned RkMatrix has rank at most rmax, and is expected to satisfy |M - R| < max(atol,rtol*|M|).

HMatrices._aca_full_pivot — Method

_aca_full_pivot(M)

Find the index of the element x ∈ M maximizing its smallest singular value. This is equivalent to minimizing the spectral norm of the inverse of x.

When x is a scalar, this is simply the element with largest absolute value.

See also: _aca_partial_pivot.

HMatrices._aca_partial — Function

_aca_partial(K,irange,jrange,atol,rmax,rtol,istart=1)

Internal function implementing the adaptive cross-approximation algorithm with partial pivoting. The returned R::RkMatrix provides an approximation to K[irange,jrange] which has either rank is expected to satisfy|M - R| < max(atol,rtol*|M|)`, but this inequality may fail to hold due to the various errors involved in estimating the error and |M|.

HMatrices._aca_partial_pivot — Method

_aca_partial_pivot(v,I)

Find in the valid set I the index of the element x ∈ v maximizing its smallest singular value. This is equivalent to minimizing the spectral norm of the inverse of x.

When x is a scalar, this is simply the element with largest absolute value.

This general implementation should work for both scalar as well as tensor-valued kernels; see (https://www.sciencedirect.com/science/article/pii/S0021999117306721)[https://www.sciencedirect.com/science/article/pii/S0021999117306721] for more details.

HMatrices._assemble_cpu! — Method

_assemble_cpu!(hmat::HMatrix,K,comp)

Assemble data on the leaves of hmat. The admissible leaves are compressed using the compressor comp. This function assumes the structure of hmat has already been intialized, and therefore should not be called directly. See HMatrix information on constructors.

HMatrices._assemble_hmat_distributed — Method

_assemble_hmat_distributed(K,rtree,ctree;adm=StrongAdmissibilityStd(),comp=PartialACA();global_index=true,threads=false)

Internal methods called after the DHMatrix structure has been initialized in order to construct the HMatrix on each of the leaves of the DHMatrix.

HMatrices._assemble_threads! — Method

_assemble_threads!(hmat::HMatrix,K,comp)

Like _assemble_cpu!, but uses threads to assemble the leaves. Note that the threads are spanwned using Threads.@spawn, which means they are spawned on the same worker as the caller.

HMatrices._build_block_structure! — Method

_build_block_structure!(adm_fun,current_node)

Recursive constructor for HMatrix block structure. Should not be called directly.

HMatrices._cost_gemv — Method

_cost_gemv(A::Union{Matrix,SubArray,Adjoint})

A proxy for the computational cost of a matrix/vector product.

HMatrices._hgemv_recursive! — Method

_hgemv_recursive!(C,A,B,offset)

Internal function used to compute C[I] <-- C[I] + A*B[J] where I = rowrange(A) - offset[1] and J = rowrange(B) - offset[2].

The offset argument is used on the caller side to signal if the original hierarchical matrix had a pivot other than (1,1).

HMatrices._update_frob_norm — Method

_update_frob_norm(acc,A,B)

Given the Frobenius norm of Rₖ = A[1:end-1]*adjoint(B[1:end-1]) in acc, compute the Frobenius norm of Rₖ₊₁ = A*adjoint(B) efficiently.

HMatrices.assemble_hmat — Method

assemble_hmat(K,rowtree,coltree;adm=StrongAdmissibilityStd(),comp=PartialACA(),threads=true,distributed=false,global_index=true)
assemble_hmat(K::KernelMatrix;threads=true,distributed=false,global_index=true,[rtol],[atol],[rank])

Main routine for assembling a hierarchical matrix. The argument K represents the matrix to be approximated, rowtree and coltree are tree structure partitioning the row and column indices, respectively, adm can be called on a node of rowtree and a node of coltree to determine if the block is compressible, and comp is a function/functor which can compress admissible blocks.

It is assumed that K supports getindex(K,i,j), and that comp can be called as comp(K,irange::UnitRange,jrange::UnitRange) to produce a compressed version of K[irange,jrange] in the form of an RkMatrix.

HMatrices.build_sequence_partition — Method

build_sequence_partition(seq,nq,cost,nmax)

Partition the sequence seq into nq contiguous subsequences with a maximum of cost of nmax per set. Note that if nmax is too small, this may not be possible (see has_partition).

HMatrices.compress! — Method

compress!(M::RkMatrix,tsvd::TSVD)

Recompress the matrix R using a truncated svd of R. The implementation uses the qr-svd strategy to efficiently compute svd(R) when rank(R) ≪ min(size(R)).

HMatrices.compress! — Method

compress!(M::Matrix,tsvd::TSVD)

Recompress the matrix M using a truncated svd and output an RkMatrix. The data in M is invalidated in the process.

HMatrices.compression_ratio — Method

compression_ratio(R::RkMatrix)

The ratio of the uncompressed size of R to its compressed size in outer product format.

HMatrices.compression_ratio — Method

compression_ratio(H::HMatrix)

The ratio of the uncompressed size of H to its compressed size.

HMatrices.disable_getindex — Method

disable_getindex()

Call this function to disable the getindex method on AbstractHMatrix. This is useful to avoid performance pitfalls associated with linear algebra methods falling back to a generic implementation which uses the getindex method. Calling getindex(H,i,j) will error after calling this function.

HMatrices.enable_getindex — Method

enable_getindex()

The opposite of disable_getindex.

HMatrices.find_optimal_cost — Function

find_optimal_cost(seq,nq,cost,tol)

Find an approximation to the cost of an optimal partitioning of seq into nq contiguous segments. The optimal cost is the smallest number cmax such that has_partition(seq,nq,cost,cmax) returns true.

HMatrices.find_optimal_partition — Function

find_optimal_partition(seq,nq,cost,tol=1)

Find an approximation to the optimal partition seq into nq contiguous segments according to the cost function. The optimal partition is the one which minimizes the maximum cost over all possible partitions of seq into nq segments.

The generated partition is optimal up to a tolerance tol; for integer valued cost, setting tol=1 means the partition is optimal.

HMatrices.flush_to_children! — Method

flush_to_children!(H::HMatrix,compressor)

Transfer the blocks data to its children. At the end, set H.data to nothing.

HMatrices.flush_to_leaves! — Method

flush_to_leaves!(H::HMatrix,compressor)

Transfer the blocks data to its leaves. At the end, set H.data to nothing.

See also: flush_to_children!

HMatrices.getcol! — Method

getcol!(col,M::AbstractMatrix,j)

Fill the entries of col with column j of M.

HMatrices.getcol — Method

getcol(M::AbstractMatrix,j)

Return a vector containing the j-th column of M.

HMatrices.has_partition — Function

has_partition(v,np,cost,cmax)

Given a vector v, determine whether or not a partition into np segments is possible where the cost of each partition does not exceed cmax.

HMatrices.hilbert_cartesian_to_linear — Method

hilbert_cartesian_to_linear(n,x,y)

Convert the cartesian indices x,y into a linear index d using a hilbert curve of order n. The coordinates x,y range from 0 to n-1, and the output d ranges from 0 to n^2-1.

See https://en.wikipedia.org/wiki/Hilbert_curve.

HMatrices.hilbert_linear_to_cartesian — Method

hilbert_linear_to_cartesian(n,d)

Convert the linear index 0 ≤ d ≤ n^2-1 into the cartesian coordinates 0 ≤ x < n-1 and 0 ≤ y ≤ n-1 on the Hilbert curve of order n.

See https://en.wikipedia.org/wiki/Hilbert_curve.

HMatrices.hilbert_partitioning — Method

hilbert_partitioning(H::HMatrix,np,cost)

Partiotion the leaves of H into np sequences of approximate equal cost (as determined by the cost function) while also trying to maximize the locality of each partition.

HMatrices.hmul! — Method

hmul!(C::HMatrix,A::HMatrix,B::HMatrix,a,b,compressor)

Similar to mul! : compute C <-- A*B*a + B*b, where A,B,C are hierarchical matrices and compressor is a function/functor used in the intermediate stages of the multiplication to avoid growring the rank of admissible blocks after addition is performed.

HMatrices.isclean — Method

isclean(H::HMatrix)

Return true if all leaves of H have data, and if the leaves are the only nodes containing data. This is the normal state of an ℋ-matrix, but during intermediate stages of a computation data may be associated with non-leaf nodes for convenience.

HMatrices.num_stored_elements — Method

num_stored_elements(R::RkMatrix)

The number of entries stored in the representation. Note that this is not length(R).

LinearAlgebra.axpy! — Method

axpy!(a::Number,X::Union{Matrix,RkMatrix,HMatrix},Y::Union{Matrix,RkMatrix,HMatrix})

Perform Y <-- a*X + Y in-place. Note that depending on the types of X and Y, this may require converting from/to different formats during intermediate calculations.

In the case where Y is an RkMatrix, the call axpy!(a,X,Y) should typically be followed by recompression stage to keep the rank of Y under control.

In the case where Y is an HMatrix, the call axpy!(a,X,Y) sums X to the data in the node Y (and not on the leaves). In case Y has no data, it will simply be assigned X. This means that after the call axpy(a,X,Y), the object Y is in a dirty state (see [isclean][@ref]) and usually a call to flush_to_leaves! or flush_to_children! follows.

LinearAlgebra.lu! — Method

lu!(M::HMatrix,comp)

Hierarhical LU facotrization of M, using comp to generate the compressed blocks during the multiplication routines.

LinearAlgebra.lu! — Method

lu!(M::HMatrix;atol=0,rank=typemax(Int),rtol=atol>0 ||
rank<typemax(Int) ? 0 : sqrt(eps(Float64)))

Hierarhical LU facotrization of M, using the PartialACA(;atol,rtol;rank) compressor.

LinearAlgebra.lu — Method

lu(M::HMatrix,args...;kwargs...)

Hierarchical LU factorization. See lu! for the available options.

LinearAlgebra.mul! — Function

mul!(y::AbstractVector,H::HMatrix,x::AbstractVector,a,b;global_index=true,threads=true)

Perform y <-- H*x*a + y*b in place.

HMatrices.@hprofile — Macro

@hprofile

A macro which

resets the default TimerOutputs.get_defaulttimer to zero

execute the code block

print the profiling details

This is useful as a coarse-grained profiling strategy in HMatrices to get a rough idea of where time is spent. Note that this relies on TimerOutputs annotations manually inserted in the code.